Removed most of pyobjc from requirements, fixed README errors
props @jkudish
