+++ requirements.txt
@@ -1,39 +1,3 @@
-distribute==0.7.3
-pyobjc==2.5.1
 pyobjc-core==2.5.1
-pyobjc-framework-Accounts==2.5.1
-pyobjc-framework-AddressBook==2.5.1
-pyobjc-framework-AppleScriptKit==2.5.1
-pyobjc-framework-AppleScriptObjC==2.5.1
-pyobjc-framework-Automator==2.5.1
-pyobjc-framework-CFNetwork==2.5.1
-pyobjc-framework-CalendarStore==2.5.1
 pyobjc-framework-Cocoa==2.5.1
-pyobjc-framework-Collaboration==2.5.1
-pyobjc-framework-CoreData==2.5.1
-pyobjc-framework-CoreLocation==2.5.1
-pyobjc-framework-CoreText==2.5.1
-pyobjc-framework-DictionaryServices==2.5.1
-pyobjc-framework-EventKit==2.5.1
-pyobjc-framework-ExceptionHandling==2.5.1
-pyobjc-framework-FSEvents==2.5.1
-pyobjc-framework-InputMethodKit==2.5.1
-pyobjc-framework-InstallerPlugins==2.5.1
-pyobjc-framework-InstantMessage==2.5.1
-pyobjc-framework-LatentSemanticMapping==2.5.1
-pyobjc-framework-LaunchServices==2.5.1
-pyobjc-framework-Message==2.5.1
-pyobjc-framework-PreferencePanes==2.5.1
-pyobjc-framework-PubSub==2.5.1
-pyobjc-framework-QTKit==2.5.1
-pyobjc-framework-Quartz==2.5.1
-pyobjc-framework-ScreenSaver==2.5.1
-pyobjc-framework-ScriptingBridge==2.5.1
-pyobjc-framework-SearchKit==2.5.1
-pyobjc-framework-ServerNotification==2.5.1
-pyobjc-framework-ServiceManagement==2.5.1
-pyobjc-framework-Social==2.5.1
-pyobjc-framework-SyncServices==2.5.1
-pyobjc-framework-SystemConfiguration==2.5.1
-pyobjc-framework-WebKit==2.5.1
-wsgiref==0.1.2
+distribute==0.7.3
